CAMEL-12460 Fix camel actuator endpoints to get it working with SB2
Also extracted CamelRoutesEndpointProperties as a separated class from CamelRoutesEndpoint
