+++ components/camel-spring-boot/src/main/java/org/apache/camel/spring/boot/actuate/endpoint/CamelRoutesEndpoint.java
@@ -39,19 +39,20 @@ import org.springframework.boot.actuate.endpoint.annotation.ReadOperation;
 import org.springframework.boot.actuate.endpoint.annotation.Selector;
 import org.springframework.boot.actuate.endpoint.annotation.WriteOperation;
 import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.lang.Nullable;
 
 /**
  * {@link Endpoint} to expose {@link org.apache.camel.Route} information.
  */
 @Endpoint(id = "camelroutes", enableByDefault = true)
-@ConfigurationProperties("management.endpoint.camelroutes")
 public class CamelRoutesEndpoint {
 
     private CamelContext camelContext;
-    private boolean readOnly = true;
+    private CamelRoutesEndpointProperties properties;
 
-    public CamelRoutesEndpoint(CamelContext camelContext) {
+    public CamelRoutesEndpoint(CamelContext camelContext, CamelRoutesEndpointProperties properties) {
         this.camelContext = camelContext;
+        this.properties = properties;
     }
 
     @ReadOperation
@@ -72,8 +73,8 @@ public class CamelRoutesEndpoint {
     }
 
     @WriteOperation
-    public void doWriteAction(@Selector String id, @Selector WriteAction action, TimeInfo timeInfo) {
-        if (this.isReadOnly()) {
+    public void doWriteAction(@Selector String id, @Selector WriteAction action, @Nullable TimeInfo timeInfo) {
+        if (this.properties.isReadOnly()) {
             throw new IllegalArgumentException(String.format("Read only: write action %s is not allowed", action));
         }
 
@@ -106,7 +107,7 @@ public class CamelRoutesEndpoint {
 
     @WriteOperation
     public String getRouteDump(@Selector String id) {
-        if (this.isReadOnly()) {
+        if (this.properties.isReadOnly()) {
             throw new IllegalArgumentException("Read only: route dump is not permitted in read-only mode");
         }
 
@@ -119,14 +120,6 @@ public class CamelRoutesEndpoint {
             }
         }
         return null;
-    }
-
-    public boolean isReadOnly() {
-        return readOnly;
-    }
-
-    public void setReadOnly(boolean readOnly) {
-        this.readOnly = readOnly;
     }
 
     private RouteEndpointInfo getRouteInfo(String id) {
