+++ components/camel-spring-boot/src/main/java/org/apache/camel/spring/boot/actuate/endpoint/CamelRoutesEndpointAutoConfiguration.java
@@ -18,11 +18,14 @@ package org.apache.camel.spring.boot.actuate.endpoint;
 
 import org.apache.camel.CamelContext;
 import org.apache.camel.spring.boot.CamelAutoConfiguration;
+import org.springframework.boot.actuate.autoconfigure.endpoint.EndpointAutoConfiguration;
 import org.springframework.boot.actuate.autoconfigure.endpoint.condition.ConditionalOnEnabledEndpoint;
+import org.springframework.boot.actuate.autoconfigure.endpoint.web.WebEndpointAutoConfiguration;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -30,17 +33,24 @@ import org.springframework.context.annotation.Configuration;
  * Auto configuration for the {@link CamelRoutesEndpoint}.
  */
 @Configuration
+@EnableConfigurationProperties({ CamelRoutesEndpointProperties.class })
 @ConditionalOnClass({CamelRoutesEndpoint.class})
 @ConditionalOnBean(CamelAutoConfiguration.class)
 @AutoConfigureAfter(CamelAutoConfiguration.class)
 public class CamelRoutesEndpointAutoConfiguration {
+
+    private CamelRoutesEndpointProperties properties;
+
+    public CamelRoutesEndpointAutoConfiguration(CamelRoutesEndpointProperties properties) {
+        this.properties = properties;
+    }
 
     @Bean
     @ConditionalOnClass(CamelContext.class)
     @ConditionalOnMissingBean
     @ConditionalOnEnabledEndpoint
     public CamelRoutesEndpoint camelEndpoint(CamelContext camelContext) {
-        return new CamelRoutesEndpoint(camelContext);
+        return new CamelRoutesEndpoint(camelContext, properties);
     }
 
 }
