+++ examples/camel-example-spring-boot-supervising-route-controller/src/main/resources/application.properties
@@ -24,13 +24,15 @@ logging.level.org.apache.camel.impl.SupervisingRouteController = DEBUG
 logging.level.org.apache.camel.util.backoff = DEBUG
 logging.level.sample.camel = DEBUG
 
+# expose actuator endpoint via HTTP
+management.endpoints.web.exposure.include=info,health,camelroutes
+
 management.endpoints.enabled-by-default = true
 management.endpoints.jmx.enabled = false
 management.endpoint.health.enabled = true
 
 # camel routes is by default enabled
 # so you do not have to configure below
-# management.endpoint.camelroutes.path = /camel/routes
 # management.endpoint.camelroutes.enabled = true
 
 # turn off read-only so we can stop/start the Camel routes
