+++ examples/camel-example-spring-boot/src/main/resources/application.properties
@@ -31,6 +31,9 @@ timer.period = 2000
 # add for example: &repeatCount=5 to the timer endpoint to make Camel idle
 #camel.springboot.duration-max-idle-seconds=15
 
+# expose actuator endpoint via HTTP
+management.endpoints.web.exposure.include=info,health,camelroutes
+
 # show verbose health details (/actuator/info) so you can see Camel information also
 management.endpoint.health.show-details=always
 
@@ -46,3 +49,4 @@ management.endpoint.camelroutes.read-only = true
 #logging.level.org.apache.camel.spring.boot = INFO
 #logging.level.org.apache.camel.impl = DEBUG
 #logging.level.sample.camel = DEBUG
+
